requirements.txt
torch>=2.7.0+cu118
torchvision>=0.22.0+cu118
fastapi>=0.115.12
uvicorn>=0.34.3
python-multipart>=0.0.20
Pillow>=11.0.0
numpy>=2.1.2
opencv-python>=4.11.0
requests>=2.32.3
matplotlib>=3.10.3
psutil>=7.0.0
kubernetes>=33.1.0
prometheus-client>=0.22.1
